fix: correct parameter type in Blake3.kt compress call
The compress function expects Int for blockLen parameter, but blockEnd.toLong() was being passed.
Changed to pass blockEnd directly as Int.

+++ dmvp-android/app/src/main/java/com/dmvp/app/security/Blake3.kt
@@ -114,7 +114,7 @@ object Blake3 {
             if (blockIdx == 0) blockFlags = blockFlags or CHUNK_START
             if (blockIdx == blockCount - 1) blockFlags = blockFlags or CHUNK_END
             
-            state = compress(state, block, chunkIndex, blockEnd.toLong(), blockFlags)
+            state = compress(state, block, chunkIndex, blockEnd, blockFlags)
         }
         
         return state.toByteArray()
